tri-search/requirements.txt
beautifulsoup4==4.14.3
faiss-cpu==1.8.0.post1
latex2mathml==3.79.0
lxml==6.0.2
numpy==1.24.4
open-clip-torch==2.32.0
opensearch-py==3.0.0
pillow==10.4.0
pydantic==2.10.6
python-dotenv==1.0.1
requests==2.33.1
sentence-transformers==3.2.1
torch==2.4.1
bs4==0.0.2
requests===2.33.1